move fundamentals to internal.fundamentals
