+++ izumi-reflect/izumi-reflect/src/main/scala/izumi/reflect/internal/fundamentals/collections/IzCollections.scala
@@ -16,7 +16,7 @@
  *
  */
 
-package izumi.reflect.fundamentals.collections
+package izumi.reflect.internal.fundamentals.collections
 
 import scala.language.implicitConversions
