+++ izumi-reflect/izumi-reflect/src/main/scala/izumi/reflect/internal/fundamentals/collections/package.scala
@@ -16,7 +16,7 @@
  *
  */
 
-package izumi.reflect.fundamentals
+package izumi.reflect.internal.fundamentals
 
 import scala.collection.mutable
