+++ izumi-reflect/izumi-reflect/src/main/scala_2.11/izumi/reflect/internal/fundamentals/platform/language/unused.scala
@@ -16,6 +16,6 @@
  *
  */
 
-package izumi.reflect.fundamentals.platform.language
+package izumi.reflect.internal.fundamentals.platform.language
 
 private[reflect] final class unused extends deprecated("unused", "unused")
