+++ izumi-reflect/izumi-reflect/src/main/scala_2.13/izumi/reflect/internal/fundamentals/platform/language/package.scala
@@ -1,4 +1,4 @@
-package izumi.reflect.fundamentals.platform
+package izumi.reflect.internal.fundamentals.platform
 
 package object language {
   private[reflect] type unused = scala.annotation.unused
